+++ links/gql_websocket_link/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.1
+
+- Handle exceptions.
+- Throw `WebSocketLinkParserException` and `WebSocketLinkServerException`.
+
 ## 0.1.0
 
 - port Websocket link from `graphq-flutter`